backend/requirements.txt
aiofiles==23.1.0
alembic==1.9.4
email-validator==1.3.1
fastapi==0.92.0
psycopg2-binary==2.9.5
pydantic[email]==1.10.5
python-jose==3.3.0
python-multipart==0.0.5
SQLAlchemy==2.0.4
starlette==0.25.0
uvicorn==0.20.0
celery==5.2.7
passlib==1.7.4
google-cloud-storage==2.9.0
google-cloud-pubsub==2.16.1
celery-pubsub==1.0.2